Version 2.1: add the theme name as a class name to the <body> element to ease per-theme styling
This allows on-wiki CSS pages such as [[MediaWiki:Vector.css]] to contain
styles for multiple themes, such as the default theme (=on-wiki
customizations on top of the skin's CSS) as well as any and all themes the
skin may have (e.g. Dark and DeepSea in Vector's case; for example,
DeepSea-specific CSS rules could be prefixed with .theme-deepsea).

HT Lcawte & SamanthaNguyen

Change-Id: I55397f13efcd3407e91772903fefa041b1c18e57