Remove unnecessary, overqualified element parts of id selectors
Those combinations origin from a time when these were seen as faster
to be rendered (at least dating back to 2011). Not only is this
reason no longer true, it also holds us back to start using semantic
HTML5 sectioning elements like `main` for `#content`.
In case of `#content`, we replace it with Vector and MonoBook's `.mw-body`.

Bug: T248137
Change-Id: Ida9d32d7d3943e077446524afacdadc0b502e554