Support debug mode
On debug mode (URL parameter `debug=true`) each style module is loaded
with an own <link> element. In this case the <link> element with the
theme module must be removed.

Also add a check to update the href attribute only if the value is
changed to avoid unnecessary DOM changes.

This change removes the last jQuery global selector and therefor the
eslint rule "no-jquery/no-global-selector": "off" can removed.

Change-Id: I8146b55b63ad08764b84479643733a555c5bb122