Fix the displaying of the "unsupported skin" note for skins which *do* have themes
Previously, with MonoBook both as the site default ($wgDefaultSkin) and user's personal preference, Special:Preferences was incorrectly claiming that the skin has no themes.
That is obviously false as MonoBook's (and Vector's) themes are bundled with the Theme extension itself.

The API response was correct, but the Theme#getSkinsWithThemes method was being passed an array containing (internal skin name -> likely i18ned/CamelCased name) mappings, like so:
$skins = Array(
	[anisa] => Anisa
	[aurora] => Aurora
	...
	[monobook] => MonoBook
	...
)
However, the ThemeModules extension attribute (defined in a skin's skin.json for most skins which support themes) looks like this as the $themes variable in that method:
$themes = Array(
	[bouquet] => Array(
		[0] => forgetmenot
		[1] => pinkdogwood
		[2] => tigerlily
	)

	[dusk] => Array(
		[0] => green
	)

	[monobook] => Array(
		[0] => dark
		[1] => pink
		[2] => stellarbook
	)

	[vector] => Array(
		[0] => dark
		[1] => dark-grey
		[2] => deepsea
	)
)

Obviously 'monobook' is not the same as 'MonoBook', so let's just use array_keys() to fix the bug.

Follow-up to 61506a96b9e0f8bfe203f7d474c432f9d2822b0f.

Change-Id: I9942048df298e85187b6da877b1b2aa583f66eff