Add $wgSkipThemes to allow sysadmins to disable certain skins' certain themes
Lightly tested with a few skins and themes and seems to work on any and all remotely sane, FOSS skins.
Does not currently work with ShoutWiki's forks of the old Wikia Quartz and Monaco skins.
Quartz is a trickier case, but for Monaco it *seems* (again, based on very little testing) that theme-related code needs to be removed from SkinMonaco#initPage and that's it.

$wgSkipThemes is an array that looks like this:
$wgSkipThemes = [
	'monobook' => [
		'stellarbook' => true
	],
	'vector' => [
		'dark' => true
	]
];
If you wanted to disable the StellarBook MonoBook and dark Vector themes, for example.

Bug: T220121
Change-Id: Ic24b2b90cf55ee2411ee65f15e8bfe74f6a633bc